+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Features
 
 - Canister status response extended with query statistics.
+- Add `metadata` function to ledger ICP.
+- Add optional parameters to ICP ledger `transactionFee`.
 
 # 2024.05.14-0630Z